requirements.txt
Django==5.1.4
openai
PyMuPDF==1.22.5
azureopenai
asgiref==3.8.1
sqlparse==0.5.3
Pillow==10.0.0
azure-storage-blob==12.24.0

# pdf2image==1.16.3
# gunicorn==20.1.0